Update conftest file Adding some fixtures from product and delivery_location modules.
