+++ delivery_location/tests/test_deliverylocation.py
@@ -1,35 +1,7 @@
 import pytest
 from delivery_location.models import DeliveryLocation
-from supplier.models import Supplier
 from django.core.exceptions import ValidationError
 import datetime
-
-
-@pytest.fixture
-def test_data():
-    return [
-        ('taxas', 'firstname', 'lastname', 'password', 'business name'),
-        ('ed', 'edd', 'eddy', 'jawbreaker101', 'jawbreaker inc'),
-    ]
-
-
-@pytest.fixture
-def supplier0(test_data):
-    return Supplier(user_name=test_data[0][0],
-                    first_name=test_data[0][1],
-                    last_name=test_data[0][2],
-                    password=test_data[0][3],
-                    business_name=test_data[0][4],)
-
-
-@pytest.fixture
-def delivery_location0(supplier0):
-    return DeliveryLocation(user_name=supplier0, location="Qiryat Shemona", date=datetime.date(2022, 12, 30))
-
-
-@pytest.fixture
-def delivery_location1(supplier0):
-    return DeliveryLocation(user_name=supplier0, location="Haifa", date=datetime.date(2022, 12, 31))
 
 
 class TestDeliveryLocationModel:
@@ -70,7 +42,7 @@ class TestDeliveryLocationModel:
         supplier0.save()
         delivery_location0.add_delivery_location()
         assert delivery_location0 in list(
-            DeliveryLocation.filter_by_supplier_and_location(supplier0, "Qiryat Shemona"))
+            DeliveryLocation.filter_by_supplier_and_location(supplier0, "Kiryat Shemona"))
 
     @pytest.mark.django_db()
     def test_filter_by_location_that_does_not_exist_from_supplier(self, supplier0):
